Aligned the stack, Fixed typo in folder name

+++ libc/include/string.h
@@ -40,7 +40,7 @@ void* memmove(void*, const void*, size_t);
  * 
  * @return void* 
  */
-void* memset(uint8_t*, char, uint8_t);
+void* memset(uint8_t*, char, uint64_t);
 /**
  * @brief Get length of a string (until `\0`)
  *
