+++ libc/string/memset.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <string.h>
 
-void* memset(uint8_t* bufptr, char value, uint8_t size) {
+void* memset(uint8_t* bufptr, char value, uint64_t size) {
 	unsigned char* buf = (unsigned char*) bufptr;
 	for (size_t i = 0; i < size; i++) {
